+ Updated "netDxf Documentation.chm". + Rename Attribute definition and Attribute Id to Tag as it is the name given by the dxf documentation. + (fixed) Loading dxf files with duplicate attribute definition tags in block attributes. Although AutoCad allows this behaviour this library does not. Having duplicate tags is not recommended in any way, since there will be now way to know which is the definition associated to the insert attribute. + (fixed) Load dxf files containing a ThumbnailImage section. + (fixed) Load dxf file encoding. + (fixed) Big HandleSeed hex numbers. The NumHandles has been changed to long, hopefully this will solve the problem, but I don't know which is the largest handle hex number AutoCad can generate. The BigInteger class of the Net Framework 4.0 will solve this problem but at the moment this library maintains its compatibility with Net 2.0.

+++ netDxf/Blocks/Block.cs
@@ -200,7 +200,7 @@ namespace netDxf.Blocks
         /// Some objects might consume more than one, is, for example, the case of polylines that will asign
         /// automatically a handle to its vertexes. The entity number will be converted to an hexadecimal number.
         /// </remarks>
-        internal override int AsignHandle(int entityNumber)
+        internal override long AsignHandle(long entityNumber)
         {
             entityNumber = this.record.AsignHandle(entityNumber);
             entityNumber = this.end.AsignHandle(entityNumber);
