+++ netDxf/Entities/Insert.cs
@@ -209,7 +209,7 @@ namespace netDxf.Entities
         /// Some objects might consume more than one, is, for example, the case of polylines that will asign
         /// automatically a handle to its vertexes. The entity number will be converted to an hexadecimal number.
         /// </remarks>
-        internal override int AsignHandle(int entityNumber)
+        internal override long AsignHandle(long entityNumber)
         {
             entityNumber = this.endSequence.AsignHandle(entityNumber);
             foreach (Attribute attrib in this.attributes)
